refactor(api)!: Rename `primitives` mod to `core`
