+++ .changes/api-primitives-core.md
@@ -1,0 +1,5 @@
+---
+'@tauri-apps/api': 'patch:breaking'
+---
+
+Changed former `tauri` module from `primitives` to `core`.
