+++ tooling/api/src/app.ts
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-License-Identifier: MIT
 
-import { invoke } from './primitives'
+import { invoke } from './core'
 
 /**
  * Application metadata and related APIs.
