+++ tooling/api/src/index.ts
@@ -15,11 +15,11 @@
 
 import * as app from './app'
 import * as event from './event'
-import * as primitives from './primitives'
+import * as core from './core'
 import * as window from './window'
 import * as path from './path'
 import * as dpi from './dpi'
 import * as tray from './tray'
 import * as menu from './menu'
 
-export { app, dpi, event, path, primitives, window, tray, menu }
+export { app, dpi, event, path, core, window, tray, menu }
