+++ tooling/api/src/menu/submenu.ts
@@ -11,7 +11,7 @@ import { MenuItem, type MenuItemOptions } from './menuItem'
 import { CheckMenuItem } from './checkMenuItem'
 import { IconMenuItem } from './iconMenuItem'
 import { PredefinedMenuItem } from './predefinedMenuItem'
-import { invoke } from '../primitives'
+import { invoke } from '../core'
 import { type LogicalPosition, PhysicalPosition, type Window } from '../window'
 import { type ItemKind, MenuItemBase, newMenu } from './base'
 import { type MenuOptions } from './menu'
