+++ tooling/api/src/tray.ts
@@ -4,7 +4,7 @@
 
 import type { Menu, Submenu } from './menu'
 import { Resource } from './internal'
-import { Channel, invoke } from './primitives'
+import { Channel, invoke } from './core'
 
 /**
  * Describes a tray event emitted when a tray icon is clicked
