+++ tooling/bench/tests/files_transfer/public/index.html
@@ -10,17 +10,17 @@
 
     <script>
       window.addEventListener('DOMContentLoaded', (event) => {
-        window.__TAURI__.primitives
+        window.__TAURI__.core
           .invoke('read_file')
           .then((_data) => {
             // success
-            window.__TAURI__.primitives.invoke('app_should_close', {
+            window.__TAURI__.core.invoke('app_should_close', {
               exitCode: 0
             })
           })
           .catch((_error) => {
             // error
-            window.__TAURI__.primitives.invoke('app_should_close', {
+            window.__TAURI__.core.invoke('app_should_close', {
               exitCode: 1
             })
           })
