+++ tooling/bench/tests/helloworld/public/index.html
@@ -10,7 +10,7 @@
 
     <script>
       window.addEventListener('DOMContentLoaded', (event) =>
-        window.__TAURI__.primitives.invoke('app_loaded_successfully')
+        window.__TAURI__.core.invoke('app_loaded_successfully')
       )
     </script>
   </body>
