+++ tooling/cli/templates/plugin/webview-src/index.ts
@@ -1,7 +1,7 @@
 {{#if license_header}}
 {{ license_header }}
 {{/if}}
-import { invoke } from '@tauri-apps/api/primitives'
+import { invoke } from '@tauri-apps/api/core'
 
 export async function execute() {
   await invoke('plugin:{{ plugin_name }}|execute')
